Create upper level throttle class and move specific segment throttlers into it
