+++ pinot-core/src/main/java/org/apache/pinot/core/data/manager/provider/DefaultTableDataManagerProvider.java
@@ -34,7 +34,6 @@ import org.apache.pinot.core.data.manager.realtime.RealtimeTableDataManager;
 import org.apache.pinot.segment.local.data.manager.TableDataManager;
 import org.apache.pinot.segment.local.utils.SegmentLocks;
 import org.apache.pinot.segment.local.utils.SegmentPreprocessThrottler;
-import org.apache.pinot.segment.local.utils.SegmentStarTreePreprocessThrottler;
 import org.apache.pinot.spi.config.instance.InstanceDataManagerConfig;
 import org.apache.pinot.spi.config.table.TableConfig;
 import org.apache.pinot.spi.stream.StreamConfigProperties;
@@ -51,19 +50,16 @@ public class DefaultTableDataManagerProvider implements TableDataManagerProvider
   private SegmentLocks _segmentLocks;
   private Semaphore _segmentBuildSemaphore;
   private SegmentPreprocessThrottler _segmentPreprocessThrottler;
-  private SegmentStarTreePreprocessThrottler _segmentStarTreePreprocessThrottler;
 
   @Override
   public void init(InstanceDataManagerConfig instanceDataManagerConfig, HelixManager helixManager,
-      SegmentLocks segmentLocks, @Nullable SegmentPreprocessThrottler segmentPreprocessThrottler,
-      @Nullable SegmentStarTreePreprocessThrottler segmentStarTreePreprocessThrottler) {
+      SegmentLocks segmentLocks, @Nullable SegmentPreprocessThrottler segmentPreprocessThrottler) {
     _instanceDataManagerConfig = instanceDataManagerConfig;
     _helixManager = helixManager;
     _segmentLocks = segmentLocks;
     int maxParallelSegmentBuilds = instanceDataManagerConfig.getMaxParallelSegmentBuilds();
     _segmentBuildSemaphore = maxParallelSegmentBuilds > 0 ? new Semaphore(maxParallelSegmentBuilds, true) : null;
     _segmentPreprocessThrottler = segmentPreprocessThrottler;
-    _segmentStarTreePreprocessThrottler = segmentStarTreePreprocessThrottler;
   }
 
   @Override
@@ -93,7 +89,7 @@ public class DefaultTableDataManagerProvider implements TableDataManagerProvider
         throw new IllegalStateException();
     }
     tableDataManager.init(_instanceDataManagerConfig, _helixManager, _segmentLocks, tableConfig, segmentPreloadExecutor,
-        errorCache, _segmentPreprocessThrottler, _segmentStarTreePreprocessThrottler);
+        errorCache, _segmentPreprocessThrottler);
     return tableDataManager;
   }
 }
