+++ pinot-core/src/test/java/org/apache/pinot/core/data/manager/BaseTableDataManagerNeedRefreshTest.java
@@ -183,8 +183,7 @@ public class BaseTableDataManagerNeedRefreshTest {
 
     IndexLoadingConfig indexLoadingConfig = new IndexLoadingConfig(tableConfig, schema);
     ImmutableSegment immutableSegment = ImmutableSegmentLoader.load(indexDir, indexLoadingConfig,
-        BaseTableDataManagerTest.SEGMENT_PREPROCESS_THROTTLER,
-        BaseTableDataManagerTest.SEGMENT_STARTREE_PREPROCESS_THROTTLER);
+        BaseTableDataManagerTest.SEGMENT_PREPROCESS_THROTTLER);
     when(segmentDataManager.getSegment()).thenReturn(immutableSegment);
     return segmentDataManager;
   }
