+++ pinot-core/src/test/java/org/apache/pinot/core/data/manager/offline/DimensionTableDataManagerTest.java
@@ -40,6 +40,7 @@ import org.apache.pinot.segment.local.segment.creator.SegmentTestUtils;
 import org.apache.pinot.segment.local.segment.creator.impl.SegmentCreationDriverFactory;
 import org.apache.pinot.segment.local.segment.index.loader.IndexLoadingConfig;
 import org.apache.pinot.segment.local.segment.index.loader.LoaderTest;
+import org.apache.pinot.segment.local.utils.SegmentAllIndexPreprocessThrottler;
 import org.apache.pinot.segment.local.utils.SegmentLocks;
 import org.apache.pinot.segment.local.utils.SegmentPreprocessThrottler;
 import org.apache.pinot.segment.local.utils.SegmentStarTreePreprocessThrottler;
@@ -78,10 +79,8 @@ public class DimensionTableDataManagerTest {
   private static final String CSV_DATA_PATH = "data/dimBaseballTeams.csv";
   private static final String SCHEMA_PATH = "data/dimBaseballTeams_schema.json";
   private static final String TABLE_CONFIG_PATH = "data/dimBaseballTeams_config.json";
-  private static final SegmentPreprocessThrottler SEGMENT_PREPROCESS_THROTTLER = new SegmentPreprocessThrottler(1, 2,
-      true);
-  private static final SegmentStarTreePreprocessThrottler SEGMENT_STARTREE_PREPROCESS_THROTTLER =
-      new SegmentStarTreePreprocessThrottler(1);
+  private static final SegmentPreprocessThrottler SEGMENT_PREPROCESS_THROTTLER = new SegmentPreprocessThrottler(
+      new SegmentAllIndexPreprocessThrottler(1, 2, true), new SegmentStarTreePreprocessThrottler(1));
 
   private File _indexDir;
   private IndexLoadingConfig _indexLoadingConfig;
@@ -153,7 +152,7 @@ public class DimensionTableDataManagerTest {
     DimensionTableDataManager tableDataManager =
         DimensionTableDataManager.createInstanceByTableName(OFFLINE_TABLE_NAME);
     tableDataManager.init(instanceDataManagerConfig, helixManager, new SegmentLocks(), tableConfig, null, null,
-        SEGMENT_PREPROCESS_THROTTLER, SEGMENT_STARTREE_PREPROCESS_THROTTLER);
+        SEGMENT_PREPROCESS_THROTTLER);
     tableDataManager.start();
     return tableDataManager;
   }
@@ -176,7 +175,7 @@ public class DimensionTableDataManagerTest {
 
     // assert that segments are released after loading data
     tableDataManager.addSegment(ImmutableSegmentLoader.load(_indexDir, _indexLoadingConfig,
-        SEGMENT_PREPROCESS_THROTTLER, SEGMENT_STARTREE_PREPROCESS_THROTTLER));
+        SEGMENT_PREPROCESS_THROTTLER));
     for (SegmentDataManager segmentManager : returnedManager.acquireAllSegments()) {
       assertEquals(segmentManager.getReferenceCount() - 1, // Subtract this acquisition
           1, // Default ref count
@@ -209,7 +208,7 @@ public class DimensionTableDataManagerTest {
     assertNull(tableDataManager.lookupValues(key, new String[]{"teamID", "teamName"}));
 
     tableDataManager.addSegment(ImmutableSegmentLoader.load(_indexDir, _indexLoadingConfig,
-        SEGMENT_PREPROCESS_THROTTLER, SEGMENT_STARTREE_PREPROCESS_THROTTLER));
+        SEGMENT_PREPROCESS_THROTTLER));
 
     // Confirm table is loaded and available for lookup
     assertTrue(tableDataManager.containsKey(key));
@@ -259,7 +258,7 @@ public class DimensionTableDataManagerTest {
     when(helixManager.getHelixPropertyStore()).thenReturn(propertyStore);
     DimensionTableDataManager tableDataManager = makeTableDataManager(helixManager);
     tableDataManager.addSegment(ImmutableSegmentLoader.load(_indexDir, _indexLoadingConfig,
-        SEGMENT_PREPROCESS_THROTTLER, SEGMENT_STARTREE_PREPROCESS_THROTTLER));
+        SEGMENT_PREPROCESS_THROTTLER));
 
     // Confirm table is loaded and available for lookup
     PrimaryKey key = new PrimaryKey(new String[]{"SF"});
@@ -317,7 +316,7 @@ public class DimensionTableDataManagerTest {
     assertNull(tableDataManager.lookupRow(key));
 
     tableDataManager.addSegment(ImmutableSegmentLoader.load(_indexDir, _indexLoadingConfig,
-        SEGMENT_PREPROCESS_THROTTLER, SEGMENT_STARTREE_PREPROCESS_THROTTLER));
+        SEGMENT_PREPROCESS_THROTTLER));
 
     // Confirm table is loaded and available for lookup
     assertTrue(tableDataManager.containsKey(key));
@@ -366,7 +365,7 @@ public class DimensionTableDataManagerTest {
 
     try {
       tableDataManager.addSegment(ImmutableSegmentLoader.load(_indexDir, _indexLoadingConfig,
-          SEGMENT_PREPROCESS_THROTTLER, SEGMENT_STARTREE_PREPROCESS_THROTTLER));
+          SEGMENT_PREPROCESS_THROTTLER));
       fail("Should error out when ErrorOnDuplicatePrimaryKey is configured to true");
     } catch (Exception e) {
       // expected;
