+++ pinot-server/src/main/java/org/apache/pinot/server/starter/ServerInstance.java
@@ -45,7 +45,6 @@ import org.apache.pinot.core.transport.InstanceRequestHandler;
 import org.apache.pinot.core.transport.QueryServer;
 import org.apache.pinot.core.transport.grpc.GrpcQueryServer;
 import org.apache.pinot.segment.local.utils.SegmentPreprocessThrottler;
-import org.apache.pinot.segment.local.utils.SegmentStarTreePreprocessThrottler;
 import org.apache.pinot.server.access.AccessControl;
 import org.apache.pinot.server.access.AccessControlFactory;
 import org.apache.pinot.server.access.AllowAllAccessFactory;
@@ -78,7 +77,6 @@ public class ServerInstance {
   private final AccessControl _accessControl;
   private final HelixManager _helixManager;
   private final SegmentPreprocessThrottler _segmentPreprocessThrottler;
-  private final SegmentStarTreePreprocessThrottler _segmentStarTreePreprocessThrottler;
 
   private final WorkerQueryServer _workerQueryServer;
   private ChannelHandler _instanceRequestHandler;
@@ -87,8 +85,7 @@ public class ServerInstance {
   private boolean _queryServerStarted = false;
 
   public ServerInstance(ServerConf serverConf, HelixManager helixManager, AccessControlFactory accessControlFactory,
-      @Nullable SegmentPreprocessThrottler segmentPreprocessThrottler,
-      @Nullable SegmentStarTreePreprocessThrottler segmentStarTreePreprocessThrottler)
+      @Nullable SegmentPreprocessThrottler segmentPreprocessThrottler)
       throws Exception {
     LOGGER.info("Initializing server instance");
     _helixManager = helixManager;
@@ -103,13 +100,12 @@ public class ServerInstance {
     ServerMetrics.register(_serverMetrics);
 
     _segmentPreprocessThrottler = segmentPreprocessThrottler;
-    _segmentStarTreePreprocessThrottler = segmentStarTreePreprocessThrottler;
 
     String instanceDataManagerClassName = serverConf.getInstanceDataManagerClassName();
     LOGGER.info("Initializing instance data manager of class: {}", instanceDataManagerClassName);
     _instanceDataManager = PluginManager.get().createInstance(instanceDataManagerClassName);
     _instanceDataManager.init(serverConf.getInstanceDataManagerConfig(), helixManager, _serverMetrics,
-        _segmentPreprocessThrottler, _segmentStarTreePreprocessThrottler);
+        _segmentPreprocessThrottler);
 
     // Initialize ServerQueryLogger and FunctionRegistry before starting the query executor
     ServerQueryLogger.init(serverConf.getQueryLogMaxRate(), serverConf.getQueryLogDroppedReportMaxRate(),
