+++ pinot-server/src/test/java/org/apache/pinot/server/api/BaseResourceTest.java
@@ -202,7 +202,7 @@ public abstract class BaseResourceTest {
     //       table config.
     TableDataManager tableDataManager = new OfflineTableDataManager();
     tableDataManager.init(instanceDataManagerConfig, mock(HelixManager.class), new SegmentLocks(), tableConfig, null,
-        null, null, null);
+        null, null);
     tableDataManager.start();
     _tableDataManagerMap.put(tableNameWithType, tableDataManager);
   }
